requirements.txt
absl-py==0.4.1
astor==0.7.1
certifi==2018.8.24
click==6.7
Flask==1.0.2
gast==0.2.0
grpcio==1.15.0
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.5
Pillow==5.2.0
protobuf==3.6.1
scipy==1.1.0
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
Werkzeug==0.14.1
